feat(core): extract portainer + auth services into core/ (Phase G)
G1: Move portainer-client, portainer-cache, portainer-normalizers, and
circuit-breaker (+ tests) to core/portainer/. Fix trace-context imports
to ../tracing/ (within core).

G2: Move settings-store, session-store, user-store, audit-logger,
event-bus, and oidc (+ tests) to core/services/. Update auth and
socket-io plugins to resolve user-store/session-store from ../services/
(within core, no more shim dependency for auth chain).

Create barrel indexes and re-export shims at old services/ paths.

Part of #706
