+++ backend/src/core/portainer/index.ts
@@ -1,0 +1,5 @@
+// Barrel export for core/portainer
+export * from './portainer-client.js';
+export * from './portainer-cache.js';
+export * from './portainer-normalizers.js';
+export * from './circuit-breaker.js';
