+++ backend/src/core/services/event-bus.test.ts
