+++ backend/src/core/services/index.ts
@@ -1,0 +1,7 @@
+// Barrel export for core/services
+export * from './settings-store.js';
+export * from './session-store.js';
+export * from './user-store.js';
+export * from './audit-logger.js';
+export * from './event-bus.js';
+export * from './oidc.js';
